+++ Assets/amusedART/Mushroom_Monster/Script/CoguLuigiScript.cs
@@ -1,18 +1,24 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
-public class CoguLuigiScript : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
+public class CoguLuigiScript : MonoBehaviour{
+
+    public Text texto;
+    [Range(0.1f,10.0f)]public float distance = 3;
+    private GameObject coguLuigi;
+    
+    void Start(){
+        texto.enabled = false;
+        coguLuigi = GameObject.FindWithTag ("Player");
     }
 
-    // Update is called once per frame
-    void Update()
-    {
-        
+    void Update(){
+        if(Vector3.Distance(transform.position,coguLuigi.transform.position) < distance){
+            texto.enabled = true;
+        } else{
+            texto.enabled = false;
+        }
     }
 }
