+++ Assets/amusedART/Mushroom_Monster/Script/CoguLuigiScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a0d9573a5199b914aa4fc009248d4d0b
+guid: b8893feeaeba23c4a8cf2ad5b4c27d8d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
