--- Assets/amusedART/Mushroom_Monster/Script/CoguMarioScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b8893feeaeba23c4a8cf2ad5b4c27d8d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
